Convert literals to size_t when using min and max. Otherwise it causes compiling error in some computers.
Contributes to CURA-4410.

+++ src/skin.cpp
@@ -450,7 +450,7 @@ void SkinInfillAreaComputation::generateInfill(SliceLayerPart& part, const Polyg
 void SkinInfillAreaComputation::generateRoofing(SliceLayerPart& part)
 {
     const size_t roofing_layer_count = mesh.settings.get<size_t>("roofing_layer_count");
-    const size_t wall_idx = std::min(2ul, mesh.settings.get<size_t>("wall_line_count"));
+    const size_t wall_idx = std::min(size_t(2), mesh.settings.get<size_t>("wall_line_count"));
 
     for (SkinPart& skin_part : part.skin_parts)
     {
